Create abstract base class and subclasses for each OS
